Added access to array of float data type.
This isn't very elegant. If one subscribes to a 
float array type, for example: 

sim/cockpit2/engine/indicators/engine_speed_rpm

the plugin returns the number of elements in the
array, followed by each of the elements. If a
tolerance is set, it is applied to every element
and if any exceeds tolerance, it'll return that 
element.

I haven't added any code to set values, and will
think some more about perhaps setting up somehow
to collect single values in the arrays.

In the code, values are stored in an array, where
the first element of the array is the number of 
elements that follow. The array is calloc'd but
never freed.